reorganize example derived nodes into a zoo submodule of obelisk_py. also made core submodule for base code.
