+++ obelisk/python/obelisk_py/core/estimation.py
@@ -4,9 +4,9 @@ from typing import Union
 import obelisk_sensor_msgs.msg as osm
 from rclpy.lifecycle.node import LifecycleState, TransitionCallbackReturn
 
-from obelisk_py.node import ObeliskNode
-from obelisk_py.obelisk_typing import ObeliskEstimatorMsg, ObeliskSensorMsg
-from obelisk_py.utils.internal import get_classes_in_module
+from obelisk_py.core.node import ObeliskNode
+from obelisk_py.core.obelisk_typing import ObeliskEstimatorMsg, ObeliskSensorMsg
+from obelisk_py.core.utils.internal import get_classes_in_module
 
 
 class ObeliskEstimator(ABC, ObeliskNode):
