+++ obelisk/python/obelisk_py/core/robot.py
@@ -3,8 +3,8 @@ from abc import ABC, abstractmethod
 import obelisk_sensor_msgs.msg as osm
 from rclpy.lifecycle.node import LifecycleState, TransitionCallbackReturn
 
-from obelisk_py.node import ObeliskNode
-from obelisk_py.obelisk_typing import ObeliskControlMsg
+from obelisk_py.core.node import ObeliskNode
+from obelisk_py.core.obelisk_typing import ObeliskControlMsg
 
 
 class ObeliskRobot(ABC, ObeliskNode):
